2.Streamlit/sqlmodel-chatbot/requirements.txt
streamlit
streamlit_authenticator
sqlmodel
python-dotenv
psycopg2